+++ docs/changelog.md
@@ -14,6 +14,7 @@ category: Administration
 * feat: ignore wrong case on attributes and multiple whitespace on import (#90)
 * feat: allow to inherit priority of the parent task explicitly (#98)
 * feat: allow to inherit reporter or use current user explicitly (#70)
+* feat: allow to inherit description of the parent task (#97)
 
 ### [23.05.0] - 2023-05-07